refactor: 회원가입 페이지 UI 분리 및 불필요 파일 정리
