--- src/features/auth/components/teacherSignUp/TeacherSignUpStatusSectionStyles.ts
@@ -1,81 +1,0 @@
-﻿import styled from '@emotion/styled';
-import { InlineButton } from '@/components/common/InlineButton';
-
-export const SuccessSection = styled.section`
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-  padding-top: 44px;
-`;
-
-export const SuccessIconContainer = styled.span`
-  width: 64px;
-  height: 64px;
-  border-radius: 999px;
-  display: inline-flex;
-  align-items: center;
-  justify-content: center;
-  background: ${({ theme }) => theme.colors.background.bg4};
-  color: ${({ theme }) => theme.colors.brand.dark};
-
-  svg {
-    width: 30px;
-    height: 30px;
-  }
-`;
-
-export const SuccessTitle = styled.h2`
-  ${({ theme }) => theme.fonts.title2};
-  margin: 34px 0 0;
-  color: ${({ theme }) => theme.colors.text.text1};
-`;
-
-export const SuccessDescription = styled.p`
-  ${({ theme }) => theme.fonts.body3};
-  margin: 14px 0 0;
-  color: ${({ theme }) => theme.colors.text.text2};
-  text-align: center;
-  line-height: 1.7;
-`;
-
-export const ClassCodeCard = styled.div`
-  width: 100%;
-  margin-top: 18px;
-  border-radius: 14px;
-  border: 1px solid ${({ theme }) => theme.colors.threadStatus.processing.border};
-  background: ${({ theme }) => theme.colors.threadStatus.processing.background};
-  padding: 18px 14px;
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-  gap: 10px;
-`;
-
-export const ClassLabel = styled.p`
-  ${({ theme }) => theme.fonts.caption};
-  margin: 0;
-  color: ${({ theme }) => theme.colors.brand.dark};
-`;
-
-export const ClassCode = styled.p`
-  ${({ theme }) => theme.fonts.title2};
-  margin: 0;
-  letter-spacing: 0.02em;
-  color: ${({ theme }) => theme.colors.text.text1};
-`;
-
-export const CopyButton = styled(InlineButton)`
-  margin-top: 16px;
-  width: 100%;
-
-  svg {
-    width: 18px;
-    height: 18px;
-  }
-`;
-
-export const PrimaryButton = styled(InlineButton)`
-  margin-top: 8px;
-  width: 100%;
-  height: 52px;
-`;
